refactor: Separação das rotas para melhor manutentabilidade e leitura
