+++ src/app.js
@@ -1,27 +1,29 @@
-import express from 'express';
-import bodyParser from 'body-parser';
-import routes from './routes/routes';
-import cors from 'cors';
+import express from 'express'
+import bodyParser from 'body-parser'
+import routes from './routes'
+import cors from 'cors'
 
-import './database';
+import './database'
 
 class App {
   constructor() {
-    this.server = express();
-    this.middlewares();
-    this.routes();
+    this.server = express()
+    this.middlewares()
+    this.routes()
   }
 
   // Registro dos middlewares da aplicação
   middlewares() {
-    this.server.use(bodyParser.json());
-    this.server.use(cors());
+    this.server.use(bodyParser.json())
+    this.server.use(cors())
   }
 
   // Registro das rotas da aplicação
   routes() {
-    this.server.use(routes);
+    routes.map((route) => {
+      this.server.use(route)
+    })
   }
 }
 
-module.exports = new App().server;
+module.exports = new App().server
