+++ src/routes/index.js
@@ -1,0 +1,27 @@
+import AuthRoutes from '../routes/authRoutes'
+import UsersRoutes from '../routes/usersRoutes'
+import ProfileImagesRoutes from '../routes/profileImagesRoutes'
+import PostsRoutes from '../routes/postsRoutes'
+import FriendshipsRoutes from '../routes/friendshipRoutes'
+import GroupsRoutes from '../routes/groupsRoutes'
+import FeedRoutes from '../routes/feedRoutes'
+import DonatioRoutes from '../routes/donationRoutes'
+import FollowGroupsRoutes from '../routes/followGroupsRoutes'
+import ChatRoutes from '../routes/chatRoutes'
+import MetaTagsRoutes from '../routes/metaTagsRoutes'
+
+const routes = [
+  AuthRoutes,
+  UsersRoutes,
+  ProfileImagesRoutes,
+  PostsRoutes,
+  FriendshipsRoutes,
+  GroupsRoutes,
+  FeedRoutes,
+  DonatioRoutes,
+  FollowGroupsRoutes,
+  ChatRoutes,
+  MetaTagsRoutes,
+]
+
+export default routes
